Git status shows all untracked files
In order to show all untracked files in a directory and not just the
directory itself, the '-uall' option is specified for 'git status.'
If a directory only contained untracked files only the directory itself
would be show without the option.

This behavior can also be set using the 'status.showUntrackedFiles'
option in the gitconfig.